Updating to revision delivered to my-Experiment team.
This version includes the Topics Analyzer module which will clean up very similar topics in subsequent runs. It includes some changes to the db tables to support this module. 

In addition, this commit removes some files that had been used for per-tag feedback which was not delivered to the myexperiment team.